Renamed one pakage and generated getters and setters
